requirements.txt
pydantic
python-dotenv
google-genai
fastapi
uvicorn